requirements.txt
absl-py==1.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
crcmod==1.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
Cython==3.0.11
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
eval-type-backport==0.2.2
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.14.0
fonttools==4.33.3
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.11.0
huggingface-hub==0.27.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mmcv==1.3.10
mmcv-full==1.7.2
# Editable install with no version control (mmdet==2.13.0)
-e /root/.local/lib/python3.8/site-packages
mmengine==0.10.5
model-index==0.1.11
MultiScaleDeformableAttention==1.0
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==4.3.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.18.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.2
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
supervisor==4.2.4
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.15.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.12
tinycss2==1.1.1
tomli==2.2.1
torch @ http://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp38-cp38-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu113/torchvision-0.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl
tornado==6.1
tqdm==4.65.2
traitlets==5.3.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wandb==0.19.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
yapf==0.43.0
zipp==3.8.0
